Given a token as evaluate, the position of "evaluate" component in the token "evaluate" could be regarded as either the first component or the last component.
The origin implementation regard it as the last one, which cause a bug
of wrong matching. For example, when matching "evaluate*" and
"evaluate", I will match * with the evaluate.

Therefore, I try to consider the component in a single-component token
as the first. At least it works for current case.